pyproject.toml
[project]
name = "yurtland"
version = "0.1.0"
description = ""
authors = [
    {name = "Dastan",email = "dastan.kubanuulu1@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django (>=5.2,<6.0)",
    "pillow (>=11.2.1,<12.0.0)",
    "python-decouple (>=3.8,<4.0)",
    "django-jazzmin (>=3.0.1,<4.0.0)",
    "django-filter (>=25.1,<26.0)",
    "django-ckeditor (>=6.7.2,<7.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
